==> Code/Curation/5.01_explore_symptoms.Rmd <==
---
title: "Step 5.01: Symptom Outcome Exploration"
author: "Mia Charifson"
date: "2024-02-12"
output: 
  html_document:
    code_folding: hide
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
knitr::opts_knit$set(root.dir = '/Users/miacharifson/Library/CloudStorage/OneDrive-NYULangoneHealth/Charifson Dissertation/')

library(tidyverse)
library(ggplot2)
library(lubridate)
library(tableone)
library(arsenal)
library(readxl)
library(ggpubr)
library(corrplot)
library(data.table)
library(forcats)
library(ggpubr)
```

## Introduction

The goal of this document is to: 

1. define the timing of the first symptom related to UF and Endo symptomology for each patient (never, before entry, at baseline, during follow-up)
2. explore the UF and Endo symptomolgy in our cohort population (which symptom, how many times)

```{r load}
pat_cohort <- read.csv('Data/Derived/pat_cohort_w_censoring.csv')
ever_proc_info <- read.csv('Data/Derived/Chapter 1/cohort_w_uf.csv')
symptoms1 <- read.csv('Data/Epic/table3_dx_symptoms1.csv')
symptoms2 <- read.csv('Data/Epic/table3_dx_symptoms2.csv')
uf_hints <- read.csv('Data/Derived/uf_all_hints.csv')
endo_hints <- read.csv('Data/Derived/endo_all_hints.csv')
address <- read.csv('Data/Derived/Chapter 1/pat_cohort_w_address.csv')
pat_cohort_new <- read.csv('Data/Derived/final_cohort_w_outcome.csv')
```

```{r join_outcomes}
pat_cohort_all_cases <- pat_cohort_new %>%
  mutate(final_outcome_status_all = case_when((uf_icd_confirmed & endo_icd_confirmed) & 
                                            uf_first_dx_date < endo_first_dx_date~'Both, UF first',
                                          (uf_icd_confirmed & endo_icd_confirmed) & 
                                            uf_first_dx_date > endo_first_dx_date~'Both, Endo first',
                                          (uf_icd_confirmed & endo_icd_confirmed) & 
                                            uf_first_dx_date == endo_first_dx_date~'Both, same day',
                                          (uf_icd_confirmed & !endo_icd_confirmed)~'UF only',
                                          (!uf_icd_confirmed & endo_icd_confirmed)~'Endo only',
                                          (!uf_icd_confirmed & !endo_icd_confirmed)~'Control')) %>%
  mutate(first_outcome_dx_date_all = case_when(final_outcome_status_all %in% 
                                             c('UF only', 'Both, UF first', 'Both, same day')~uf_first_dx_date,
                                           final_outcome_status_all %in% 
                                             c('Endo only', 'Both, Endo first')~endo_first_dx_date),
         outcome_before_censor = case_when(final_outcome_status_all == 'Control'~NA,
                                           first_outcome_dx_date_all <= final_censor_date~TRUE,
                                           TRUE~FALSE),
         final_censor_type_all = case_when(final_outcome_status_all != 'Control' & outcome_before_censor~'Case',
                                               final_outcome_status_all %in% 
                                                 c('UF only', 'Both, UF first', 'Both, same day') &
                                                 final_censor_type==uf_first_confirm_type~'Case',
                                               final_outcome_status_all %in% c('Endo only', 'Both, Endo first') &
                                                 final_censor_type==endo_first_confirm_type~'Case',
                                               TRUE~final_censor_type),
         final_censor_date_all = case_when(final_censor_type_all=='Case'~as.Date(first_outcome_dx_date_all),
                                               TRUE~as.Date(final_censor_date))) 
```

```{r join_symptoms}
symptoms <- bind_rows(symptoms1, symptoms2) %>%
  inner_join(pat_cohort %>% select(pat_id, index_enc_date), by='pat_id') %>%
  filter(as.Date(dx_date)<='2023-08-01')
```

## First, some cleaning

There are `r length(which(symptoms$dx_date == 'NULL'))` entries where the date of the diagnosis code is missing. All of these are `Medical History` type diagnoses. These are generally uninformative without a diagnosis date. 

TO DO: try to recover diagnosis dates for as many of these rows as possible, by joining on encounter ID.

```{r, eval=FALSE}
table(symptoms$dx_type[which(symptoms$dx_date=='NULL')])

symptoms %>%
  group_by(pat_id) %>%
  filter(n() == sum(dx_type == 'Medical History')) %>%
  arrange(pat_id)
```

## First symptom descriptives {.tabset}

There are `r length(unique(symptoms$pat_id))` patients in our cohort who have at least one UF or Endo related symptom in their EHR. Since back pain and diarhhea are relatively non-specific symptoms, we only count them as the first symptom if they are later followed by another UF/Endo related symptom.

```{r first_symptom}
first_symptom <- symptoms %>%
  group_by(pat_id) %>%
  summarise(index_enc_date = min(index_enc_date),
            first_dx_date = min(dx_date, na.rm=TRUE),
            n_specific_symptoms = sum(!grepl('M54.5|R19.7', icd10)),
            .groups='drop') %>%
  mutate(id = paste0(pat_id, first_dx_date),
         first_dx_date = as.Date(first_dx_date), 
         index_enc_date = as.Date(index_enc_date), 
         time_group = case_when(first_dx_date<index_enc_date~'before baseline', 
                                first_dx_date>=index_enc_date&first_dx_date<=(ymd(index_enc_date)+years(1))~'at baseline',
                                first_dx_date>(ymd(index_enc_date)+years(1))~'during follow-up',
                                TRUE~'has symptom, no date'))

count <- symptoms %>%
  mutate(id = paste0(pat_id, dx_date)) %>%
  group_by(id) %>%
  summarise(count_per_day = n_distinct(dx_name), 
            pat_id = unique(pat_id),
            .groups = 'drop') %>%
  group_by(pat_id) %>%
  summarise(symptom_count = sum(count_per_day), .groups = 'drop')

first_symptom_types <- symptoms %>%
  mutate(id = paste0(pat_id, dx_date)) %>%
  filter(id %in% first_symptom$id) %>%
  group_by(pat_id) %>%
  summarise(first_dx_names = paste0(unique(dx_name),collapse = ', '),
            icd10s = paste0(unique(icd10),collapse = ', '),
            .groups='drop')

first_symptom <- first_symptom %>%
  left_join(first_symptom_types, by='pat_id') %>%
  left_join(count, by='pat_id') %>%
  # flag symptom case if they have back pain/diarrhea and no other specific UF/Endo symptoms
  mutate(flag = case_when(grepl('M54.5|R19.7', icd10s) & n_specific_symptoms == 0~TRUE,
                          TRUE~FALSE))

flag_ids <- unique(first_symptom$pat_id[which(first_symptom$flag)])
```

The following questions get explored: 

1. When do patients have their first UF or Endo related symptom?
2. For patients who have symptoms, how many symptoms do they report since their first one?
3. Which symptoms are commonly reported in this cohort?
4. Which symptom is commonly reported for the first symptom at baseline & during follow-up?
5. What symptoms are associated with each other?

### 1

```{r first_symptom_table, results='asis'}
pat_cohort %>%
  select(pat_id) %>%
  left_join(first_symptom %>% filter(!flag), by='pat_id') %>%
  mutate(time_group = ifelse(is.na(time_group), 'never', time_group)) %>%
  tableby(~time_group, data=.) %>%
  summary()
```

### 2

A distinct symptom is counted as a unique diagnosis date-ICD10 code combinations.

```{r symptom_count_fig}
symptoms %>%
  filter(!pat_id %in% flag_ids) %>%
  group_by(pat_id) %>%
  summarise(count_symptoms = n_distinct(dx_date, icd10), .groups='drop') %>%
  ggplot(aes(x=count_symptoms), data=.) +
  geom_histogram(fill= 'gray', color='black', bins = 100) +
  theme_bw()
```

```{r symptom_count_table, results='asis'}
symptoms %>%
  filter(!pat_id %in% flag_ids) %>%
  group_by(pat_id) %>%
  summarise(count_symptoms = n_distinct(dx_date, icd10), .groups='drop') %>%
  tableby(~count_symptoms, data=., numeric.stats = c('mean', 'median', 'q1q3', 'range')) %>%
  summary()
```

### 3

```{r symptoms_group}
symptom_wide_boolean <- symptoms %>%
  filter(!pat_id %in% flag_ids) %>%
  group_by(pat_id, dx_date) %>%
  summarise(Pelvic_pain = (sum(grepl('R10.2', icd10))>=1),
         Dysmenorrhea = (sum(grepl('N94.[4-6]', icd10))>=1),
         Dyspareunia = (sum(grepl('F52.6|N94.1', icd10))>=1),
         Infertility = (sum(grepl('N97.9', icd10))>=1),
         Intermenstrual_pain = (sum(grepl('N94.0', icd10))>=1),
         Ovarian_cysts = (sum(grepl('N83.0|N83.2', icd10))>=1),
         Anemia = (sum(grepl('D50.0', icd10))>=1),
         Back_pain = (sum(grepl('M54.5', icd10))>=1),
         Postcoital_bleed = (sum(grepl('N93.0', icd10))>=1),
         AUB = (sum(grepl('N93.8|N93.9', icd10))>=1),
         Diarrhea = (sum(grepl('R19.7', icd10))>=1),
         Menorrhagia = (sum(grepl('N92.0|N92.1', icd10))>=1),
         Bladder_rectum_symptoms = (sum(grepl('N32.9|K59.00', icd10))>=1),
         Urinary_urgency = (sum(grepl('N39.3', icd10))>=1),
         PID = (sum(grepl('N73.9', icd10))>=1),
         .groups='drop')
```

```{r common_symptoms_pat}
patient_counts <- symptom_wide_boolean %>%
  summarise(Pelvic_pain = n_distinct(pat_id[Pelvic_pain]),
         Dysmenorrhea = n_distinct(pat_id[Dysmenorrhea]),
         Dyspareunia = n_distinct(pat_id[Dyspareunia]),
         Infertility = n_distinct(pat_id[Infertility]),
         Intermenstrual_pain = n_distinct(pat_id[Intermenstrual_pain]),
         Ovarian_cysts = n_distinct(pat_id[Ovarian_cysts]),
         Anemia = n_distinct(pat_id[Anemia]),
         Back_pain = n_distinct(pat_id[Back_pain]),
         Postcoital_bleed = n_distinct(pat_id[Postcoital_bleed]),
         AUB = n_distinct(pat_id[AUB]),
         Diarrhea = n_distinct(pat_id[Diarrhea]),
         Menorrhagia = n_distinct(pat_id[Menorrhagia]),
         Bladder_rectum_symptoms = n_distinct(pat_id[Bladder_rectum_symptoms]),
         Urinary_urgency = n_distinct(pat_id[Urinary_urgency]),
         PID = n_distinct(pat_id[PID]),) %>%
  pivot_longer(cols=everything(), names_to='Symptom', values_to='Count') %>%
  mutate(Percent = paste0(round(Count/nrow(pat_cohort)*100, 2), '%'))

patient_counts %>%
  ggplot(aes(x=Count, y=fct_reorder(Symptom, Count), fill=Symptom), data=.) +
  geom_col() + 
  geom_label(aes(label=Percent), fill='white', label.padding = unit(0.1, "lines")) +
  theme_bw() + 
  ylab('Symptom Name') +
  xlab('Count of unique patients') +
  theme(legend.position = 'none')
```

```{r common_symptoms_count}
symptom_wide_boolean %>% 
  summarise(Pelvic_pain = sum(Pelvic_pain),
         Dysmenorrhea = sum(Dysmenorrhea),
         Dyspareunia = sum(Dyspareunia),
         Infertility = sum(Infertility),
         Intermenstrual_pain = sum(Intermenstrual_pain),
         Ovarian_cysts = sum(Ovarian_cysts),
         Anemia = sum(Anemia),
         Back_pain = sum(Back_pain),
         Postcoital_bleed = sum(Postcoital_bleed),
         AUB = sum(AUB),
         Diarrhea = sum(Diarrhea),
         Menorrhagia = sum(Menorrhagia),
         Bladder_rectum_symptoms = sum(Bladder_rectum_symptoms),
         Urinary_urgency = sum(Urinary_urgency),
         PID = sum(PID)) %>%
  pivot_longer(cols=everything(), names_to='Symptom', values_to='Count') %>%
  ggplot(aes(x=Count, y=fct_reorder(Symptom, Count), fill=Symptom), data=.) +
  geom_col() +
  theme_bw() + 
  ylab('Symptom Name') +
  xlab('Count of unique instances') +
  theme(legend.position = 'none')
```

### 4

```{r common_symptoms_fu}
n_timegroup <- table(first_symptom$time_group)
patient_counts <- first_symptom %>%
  filter(!flag) %>%
  group_by(time_group) %>%
  summarise(Pelvic_pain = (sum(grepl('R10.2', icd10s))),
         Dysmenorrhea = (sum(grepl('N94.[4-6]', icd10s))),
         Dyspareunia = (sum(grepl('F52.6|N94.1', icd10s))),
         Infertility = (sum(grepl('N97.9', icd10s))),
         Intermenstrual_pain = (sum(grepl('N94.0', icd10s))),
         Ovarian_cysts = (sum(grepl('N83.0|N83.2', icd10s))),
         Anemia = (sum(grepl('D50.0', icd10s))),
         Back_pain = (sum(grepl('M54.5', icd10s))),
         Postcoital_bleed = (sum(grepl('N93.0', icd10s))),
         AUB = (sum(grepl('N93.8|N93.9', icd10s))),
         Diarrhea = (sum(grepl('R19.7', icd10s))),
         Menorrhagia = (sum(grepl('N92.0|N92.1', icd10s))),
         Bladder_rectum_symptoms = (sum(grepl('N32.9|K59.00', icd10s))),
         Urinary_urgency = (sum(grepl('N39.3', icd10s))),
         PID = (sum(grepl('N73.9', icd10s)))) %>%
  pivot_longer(cols=c(Pelvic_pain:PID), names_to='Symptom', values_to='Count') %>%
  mutate(Percent = case_when(time_group == 'at baseline' ~ Count/n_timegroup[1],
                             time_group == 'during follow-up' ~ Count/n_timegroup[3],
                             time_group == 'before baseline' ~ Count/n_timegroup[2]), 
         Percent = paste0(round(Percent*100, 2), '%'))

patient_counts %>%
  filter(time_group %in% c('at baseline', 'during follow-up')) %>%
  ggplot(aes(x=Count, y=fct_reorder(Symptom, Count), fill=Symptom), data=.) +
  geom_col() + 
  geom_label(aes(label=Percent), fill='white', label.padding = unit(0.1, "lines")) +
  theme_bw() + 
  ylab('Symptom Name') +
  xlab('Count of unique patients') +
  theme(legend.position = 'none') +
  facet_wrap(~time_group)
```

### 5

```{r symptoms_wide1}
symptom_wide <- symptoms %>%
  filter(!pat_id %in% flag_ids) %>%
  group_by(pat_id, dx_date) %>%
  summarise(Pelvic_pain = (sum(grepl('R10.2', icd10))>=1),
         Dysmenorrhea = (sum(grepl('N94.[4-6]', icd10))>=1),
         Dyspareunia = (sum(grepl('F52.6|N94.1', icd10))>=1),
         Infertility = (sum(grepl('N97.9', icd10))>=1),
         Intermenstrual_pain = (sum(grepl('N94.0', icd10))>=1),
         Ovarian_cysts = (sum(grepl('N83.0|N83.2', icd10))>=1),
         Anemia = (sum(grepl('D50.0', icd10))>=1),
         Back_pain = (sum(grepl('M54.5', icd10))>=1),
         Post_coital_bleed = (sum(grepl('N93.0', icd10))>=1),
         AUB = (sum(grepl('N93.8|N93.9', icd10))>=1),
         Diarrhea = (sum(grepl('R19.7', icd10))>=1),
         Menorrhagia = (sum(grepl('N92.0|N92.1', icd10))>=1),
         Bladder_rectum_symptoms = (sum(grepl('N32.9|K59.00', icd10))>=1),
         Urinary_urgency = (sum(grepl('N39.3', icd10))>=1),
         PID = (sum(grepl('N73.9', icd10))>=1),
         .groups='drop')
```

```{r symptoms_wide2}
symptom_wide_count <- symptoms %>%
  filter(!pat_id %in% flag_ids) %>%
  group_by(pat_id) %>%
  summarise(Pelvic_pain = (sum(grepl('R10.2', icd10))),
         Dysmenorrhea = (sum(grepl('N94.[4-6]', icd10))),
         Dyspareunia = (sum(grepl('F52.6|N94.1', icd10))),
         Infertility = (sum(grepl('N97.9', icd10))),
         Intermenstrual_pain = (sum(grepl('N94.0', icd10))),
         Ovarian_cysts = (sum(grepl('N83.0|N83.2', icd10))),
         Anemia = (sum(grepl('D50.0', icd10))),
         Back_pain = (sum(grepl('M54.5', icd10))),
         Post_coital_bleed = (sum(grepl('N93.0', icd10))),
         AUB = (sum(grepl('N93.8|N93.9', icd10))),
         Diarrhea = (sum(grepl('R19.7', icd10))),
         Menorrhagia = (sum(grepl('N92.0|N92.1', icd10))),
         Bladder_rectum_symptoms = (sum(grepl('N32.9|K59.00', icd10))),
         Urinary_urgency = (sum(grepl('N39.3', icd10))),
         PID = (sum(grepl('N73.9', icd10))),
         .groups='drop')
```

```{r corrplot}
corrplot.mixed(cor(symptom_wide_count[, -c(1)]), upper = 'square')
```

```{r heatmap}
require(graphics); require(grDevices)
x <- cor(symptom_wide_count[, -c(1)])
rc <- rainbow(nrow(x), start = 0, end = .3)
cc <- rainbow(ncol(x), start = 0, end = .3)
heatmap(x, RowSideColors = rc, ColSideColors = cc)
```

### {.unnumbered}

## Second symptom

```{r second_symptom}
second_symptom <- symptoms %>%
  group_by(pat_id) %>%
  mutate(rown = row_number()) %>%
  ungroup() %>%
  filter(rown==2 & dx_date >= index_enc_date) %>%
  rename(second_symptom_date = dx_date,
         second_symptom_dx_name = dx_name, 
         second_symptom_icd10s = icd10) %>%
  select(pat_id, contains('second'))

symptom_cohort <- first_symptom %>%
  filter(!flag & first_dx_date >= index_enc_date) %>%
  left_join(second_symptom, by='pat_id')
```

There are `r length(which(!is.na(symptom_cohort$second_symptom_dx_name)))` patients who have a second symptom following their first symptom. Of these, `r length(which(symptom_cohort$first_dx_date == symptom_cohort$second_symptom_date))` have their second symptom reported on the same day as their first symptom (in other words, their TTD won't change for two symptom cohort).

```{r second_symptoms_group}
second_symptom_count <- symptom_cohort %>%
  select(second_symptom_icd10s) %>%
  summarise(Pelvic_pain = sum(grepl('R10.2', second_symptom_icd10s)),
            Dysmenorrhea = sum(grepl('N94.[4-6]', second_symptom_icd10s)),
            Dyspareunia = sum(grepl('F52.6|N94.1', second_symptom_icd10s)),
            Infertility = sum(grepl('N97.9', second_symptom_icd10s)),
            Intermenstrual_pain = sum(grepl('N94.0', second_symptom_icd10s)),
            Ovarian_cysts = sum(grepl('N83.0|N83.2', second_symptom_icd10s)),
            Anemia = sum(grepl('D50.0', second_symptom_icd10s)),
            Back_pain = sum(grepl('M54.5', second_symptom_icd10s)),
            Postcoital_bleed = sum(grepl('N93.0', second_symptom_icd10s)),
            AUB = sum(grepl('N93.8|N93.9', second_symptom_icd10s)),
            Diarrhea = sum(grepl('R19.7', second_symptom_icd10s)),
            Menorrhagia = sum(grepl('N92.0|N92.1', second_symptom_icd10s)),
            Bladder_rectum_symptoms = sum(grepl('N32.9|K59.00', second_symptom_icd10s)),
            Urinary_urgency = sum(grepl('N39.3', second_symptom_icd10s)),
            PID = sum(grepl('N73.9', second_symptom_icd10s))) %>%
  pivot_longer(cols=everything(), names_to='Symptom', values_to='Count') %>%
  mutate(Percent = paste0(round(Count/nrow(symptom_cohort[which(!is.na(symptom_cohort$second_symptom_date)),])*100, 2), '%'))

second_symptom_count %>%
  ggplot(aes(x=Count, y=fct_reorder(Symptom, Count), fill=Symptom), data=.) +
  geom_col() + 
  geom_label(aes(label=Percent), fill='white', label.padding = unit(0.1, "lines")) +
  theme_bw() + 
  ylab('Symptom Name') +
  xlab('Count of unique patients') +
  theme(legend.position = 'none')
```

## Time to diagnoisis {.tabset}

```{r join_uf_cases}
pat_cohort_symptoms <- pat_cohort_new %>%
  inner_join(symptom_cohort %>%
              rename(first_symptom_date = first_dx_date, 
                     first_symptom_dx_name = first_dx_names, 
                     first_symptom_icd10s = icd10s) %>%
              select(-index_enc_date, -n_specific_symptoms, -time_group, -flag),
            by='pat_id') %>%
  # make final censor date end of FU if NA
  mutate(final_censor_date = case_when(is.na(final_censor_date)~as.Date('2023-08-01'),
                                       TRUE~as.Date(final_censor_date)))

pat_cohort_symptoms_all <- pat_cohort_all_cases %>%
  inner_join(symptom_cohort %>%
              rename(first_symptom_date = first_dx_date, 
                     first_symptom_dx_name = first_dx_names, 
                     first_symptom_icd10s = icd10s) %>%
              select(-index_enc_date, -n_specific_symptoms, -time_group, -flag),
            by='pat_id') %>%
  # make final censor date end of FU if NA
  mutate(final_censor_date_all = case_when(is.na(final_censor_date_all)~as.Date('2023-08-01'),
                                       TRUE~as.Date(final_censor_date_all)))
```

There are `r length(unique(pat_cohort_symptoms$pat_id))` patients with a first symptom during follow-up. However, further exclusions must be made if the first symptom is at or after censoring/after outcome diagnosis (n=`r length(which(pat_cohort_symptoms$first_symptom_date >= pat_cohort_symptoms$final_censor_date))`).

```{r remove_after_dx_symptoms}
pat_cohort_symptoms <- pat_cohort_symptoms %>%
  filter(first_symptom_date < final_censor_date)

pat_cohort_symptoms_all <- pat_cohort_symptoms_all %>% 
  filter(first_symptom_date < final_censor_date_all)
```

The following descriptives are reported: 

1. TTD overall in cohort
2. TTD by outcome status
3. TTD by first symptom type
4. First symptom type by outcome status 

### 1

For all patients with a symptom:
```{r look_at_dates}
pat_cohort_symptoms %>%
  ggplot(data=.) +
  geom_histogram(aes(x=first_symptom_date), fill='orange', color='white', bins=30) +
  xlab('Date of first symptom') +
  theme_bw()
```

For patients with a second symptom:
(red line = first symptom date, blue line = second symptom date)
```{r second_symptom_date}
ggplot(data=symptom_cohort %>% filter(!is.na(second_symptom_date))) +
  geom_density(aes(x=as.Date(first_dx_date)), color='red') +
  geom_density(aes(x=as.Date(second_symptom_date)), color='blue') +
  xlab('Date of symptom') +
  theme_bw()
```
Summary of time between first and second symptom:
```{r symptom_gap, results='asis'}
symptom_cohort %>%
  mutate(gap = as.numeric(as.Date(second_symptom_date)-as.Date(first_dx_date))) %>%
  tableby(~gap, data=., numeric.stats = c('Nmiss', 'meansd', 'median', 'q1q3', 'range')) %>%
  summary(., labelTranslations=list('gap' = 'Time between first and second symptom (days)'))
```


### 2

```{r calculate_ttd, results='asis'}
pat_cohort_symptoms$ttd <- as.numeric(difftime(pat_cohort_symptoms$final_censor_date,
                                               pat_cohort_symptoms$first_symptom_date,
                                               unit='days'))/30

pat_cohort_symptoms$ttd_2 <- as.numeric(difftime(pat_cohort_symptoms$final_censor_date,
                                                 pat_cohort_symptoms$second_symptom_date,
                                                 unit='days'))/30

pat_cohort_symptoms$ttd_2 <- ifelse(pat_cohort_symptoms$second_symptom_date >= pat_cohort_symptoms$final_censor_date,
                                    NA, pat_cohort_symptoms$ttd_2)

ttd_labs <- list('ttd'='Time from first symptom to diagnosis (months)',
                 'ttd_2'='Time from second symptom to diagnosis (months)')

## show table for ttd by outcome status
summary(tableby(final_outcome_status~ttd+ttd_2, data=pat_cohort_symptoms), labelTranslations= ttd_labs)
```

Same descriptives but for all cases:
```{r, results='asis'}
# do the same for all cases
pat_cohort_symptoms_all$ttd <- as.numeric(difftime(pat_cohort_symptoms_all$final_censor_date_all,
                                               pat_cohort_symptoms_all$first_symptom_date,
                                               unit='days'))/30

pat_cohort_symptoms_all$ttd_2 <- as.numeric(difftime(pat_cohort_symptoms_all$final_censor_date_all,
                                                 pat_cohort_symptoms_all$second_symptom_date,
                                                 unit='days'))/30

pat_cohort_symptoms_all$ttd_2 <- ifelse(pat_cohort_symptoms_all$second_symptom_date >= pat_cohort_symptoms_all$final_censor_date_all,
                                    NA, pat_cohort_symptoms_all$ttd_2)

## show table for ttd by outcome status
summary(tableby(final_outcome_status_all~ttd+ttd_2, data=pat_cohort_symptoms_all), labelTranslations= ttd_labs)
```


### 3

```{r avg_ttd_by_symptom, results='asis'}
pat_cohort_symptoms %>%
  mutate(symptom_type = case_when(grepl('R10.2|F52.6|N94.1|N94.0|M54.5', first_symptom_icd10s)~'Pain',
                                  grepl('N94.[4-6]|N93.0|N93.8|N93.9|N92.0|N92.1|D50.0',
                                             first_symptom_icd10s)~'Menstrual/Bleeding',
                                  grepl('N83.0|N83.2|N73.9', first_symptom_icd10s)~'PID/Ovarian cysts',
                                  grepl('N97.9', first_symptom_icd10s)~'Infertility',
                                  grepl('N32.9|K59.00|N39.3|R19.7', first_symptom_icd10s)~
                                    'Other Genito-urinary')) %>%
  tableby(symptom_type~ttd, data=.) %>%
  summary(., labelTranslations= ttd_labs)
```

Pain includes pelvic pain, dyspareunia, inter-menstrual pain, and back pain. Menstrual/bleeding includes dysmenorrhea, anemia, post-coital bleed, abnormal uterine bleeding, and menorrhagia. Infertility just includes infertility. Other genito-urinary symptoms include diarrhea, bladder and rectum symptoms, and urinary urgency. PID and ovarian cysts have their own category as they are related diagnoses which may be misdiagnoses or intermediary diagnoses that then warrant further screening. 

### 4

```{r symptom_by_case, results='asis'}
pat_cohort_symptoms %>%
  mutate(symptom_type = case_when(grepl('R10.2|F52.6|N94.1|N94.0|M54.5', first_symptom_icd10s)~'Pain',
                                  grepl('N94.[4-6]|N93.0|N93.8|N93.9|N92.0|N92.1|D50.0',
                                             first_symptom_icd10s)~'Menstrual/Bleeding',
                                  grepl('N83.0|N83.2|N73.9', first_symptom_icd10s)~'PID/Ovarian cysts',
                                  grepl('N97.9', first_symptom_icd10s)~'Infertility',
                                  grepl('N32.9|K59.00|N39.3|R19.7', first_symptom_icd10s)~
                                    'Other Genito-urinary')) %>%
  tableby(final_outcome_status~symptom_type, data=.) %>%
  summary(., labelTranslations= list('final_outcome_status'='Patient outcome group'))
```

### {.unnumbered}

## Identifying potential false negatives

False negatives are defined as patients with at least two distinct symptoms related to UF and/or Endo with no UF and Endo diagnostic codes. Below I will identify and flag these patients and also separate them into those with and without diagnostic procedures.

I leave off back pain and diarrhea off this as they are non-specific symptoms. I allow symptoms to occur anytime prior to the final censoring date (including case diagnosis) since for identifying false negatives, we do not care if their symptoms occur prior to baseline.

```{r false_negatives}
false_negatives <- symptoms %>%
  inner_join(pat_cohort_new %>% 
               select(pat_id, final_censor_date), 
            by='pat_id') %>%
  filter(dx_date <= final_censor_date) %>%
  group_by(pat_id) %>%
  summarise(n_specific_symptoms = sum(!grepl('M54.5|R19.7', icd10)),
            .groups='drop') %>%
  filter(n_specific_symptoms > 1) %>%
  left_join(ever_proc_info %>% select(pat_id, ever_imaging, ever_surgery), by='pat_id') %>%
  filter(!pat_id %in% pat_cohort_new$pat_id[which(pat_cohort_new$final_censor_type == 'Case')])
```

There are `r length(unique(false_negatives$pat_id))` patients who report at least two symptoms related to UF/Endo but never receive a UF or Endo diagnosis. Of these, 
there are `r length(unique(false_negatives$pat_id[which(!false_negatives$ever_imaging & !false_negatives$ever_surgery)]))` patients who additionally never received any  imaging or surgery procedures commonly used to diagnose UF/Endo during follow-up . This amounts to about `r round(length(unique(false_negatives$pat_id[which(!false_negatives$ever_imaging & !false_negatives$ever_surgery)]))/dim(pat_cohort)[1]*100, 2)`% of patients being qualified as potential false negatives. 

```{r false_negatives_graph}
p1 <- ggplot(data=false_negatives, aes(y=n_specific_symptoms, x=ever_imaging)) +
  geom_boxplot(fill='lightyellow') +
  xlab('Ever had imaging during FU') + 
  ylab('Number of symptoms reported') +
  theme_bw()

p2 <- ggplot(data=false_negatives, aes(x=ever_surgery, y=n_specific_symptoms)) +
  geom_boxplot(fill='lightyellow') +
  xlab('Ever had surgery during FU') + 
  ylab('Number of symptoms reported') +
  theme_bw()

ggarrange(p1, p2)
```

## Write output files

```{r pat_all_symptoms}
pat_symptoms_all <- pat_cohort %>%
  select(pat_id) %>%
  left_join(first_symptom %>% filter(!flag) %>% select(pat_id, time_group, symptom_count), by='pat_id') %>%
  left_join(symptom_cohort %>% select(pat_id, second_symptom_date), by='pat_id') %>%
  mutate(time_group = ifelse(is.na(time_group), 'never', time_group),
         first_symptom = case_when(time_group != 'never'~TRUE, TRUE~FALSE),
         second_symptom = case_when(!is.na(second_symptom_date)~TRUE, TRUE~FALSE)) %>%
  select(-second_symptom_date)
```


```{r save_files}
# write out false negatives and chapter 2 cohort
write.csv(pat_symptoms_all, 'Data/Derived/Chapter 2/pat_cohort_symptom_status.csv', row.names = FALSE)
write.csv(pat_cohort_symptoms, 'Data/Derived/Chapter 2/pat_cohort_w_ttd.csv', row.names = FALSE)
write.csv(pat_cohort_symptoms_all, 'Data/Derived/Chapter 2/pat_cohort_w_ttd_all_cases.csv', row.names = FALSE)
write.csv(false_negatives, 'Data/Derived/putative_false_negatives.csv', row.names = FALSE)
```
